Use Agg matplot backend if no display is found
The TkAgg backend does not work in an environment which has no graphical
user interface. It always fails. This makes it impossible to run this
tool inside a docker container for example.

The Agg backend on the other hand supports running in those
environments.

The code change makes matplot use the Agg backend when it detects that the
TkAgg backend would fail.

Signed-off-by: Niklas Wimmer <nwimmer@student.tgm.ac.at>